make rust-analyzer happy
